Organization rework + [WIP]  new boundary label notation for 1D + tests
