+++ dgtd/integrators/CN.py
@@ -1,7 +1,7 @@
 import numpy as np
 from scipy.optimize import fsolve
 
-from .spatialDiscretization import *
+from dgtd.spatialDiscretization import *
 #Crank Nicolson method
 
 class CN:
