+++ dgtd/integrators/DIRK2.py
@@ -1,7 +1,7 @@
 import numpy as np
 from scipy.optimize import fsolve
 
-from .spatialDiscretization import *
+from dgtd.spatialDiscretization import *
 #RK4
 
 
@@ -9,24 +9,26 @@ from .spatialDiscretization import *
 # b = np.array([1/2,              1/2])
 # c = np.array([1/2-np.sqrt(3)/6, 1/2+np.sqrt(3)/6])
 
-class IGLRK4:
+class DIRK2:
     def __init__(self, sp: SpatialDiscretization, fields):
         self.sp = sp
         self.time = 0.0
 
         self.A = sp.buildEvolutionOperator()          
 
-    def k_residual(self, k, f, dt, yo):
-        #k = k.reshape(2, len(yo), order='F')
-        # k1 = k[:len(yo)]
-        # k2 = k[len(yo):]
-        # k = np.vstack((k1, k2))
-        k1, k2 = k
-        return np.array([k1 - np.matmul(f, yo + dt/4 *k1 + (1/4 - np.sqrt(3)/6)*dt*k2), k2 - np.matmul(f, yo + (1/4 + np.sqrt(3)/6)*dt*k1 + 1/4*dt*k2) ])
+    def k1_residual(self, k1, f, dt, yo):
+        return k1 - np.matmul(f, yo + dt/4 *k1)
+    
+    def k2_residual(self, k2, f, dt, y1):
+        return k2-np.matmul(f, y1 + dt/4*k2)
 
     def step(self, fields, dt):
         yo = self.sp.convertToVector(fields)
-        k1, k2 = fsolve(self.k_residual, (np.ones(yo), np.ones(yo)), args = (self.A, dt, yo ))
+        k1 = np.ones(yo.shape)
+        k2 = np.ones(yo.shape)
+        k1 = fsolve(self.k1_residual, k1, args = (self.A, dt, yo ))
+        y1 = yo + dt/2*k1
+        k2 = fsolve(self.k2_residual, k2, args = (self.A, dt, y1 ))
         yp = yo + dt/2 * (k1 + k2)
         
         self.sp.copyVectorToFields(yp, fields)
