+++ dgtd/integrators/LSERK134.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from .spatialDiscretization import *
+from dgtd.spatialDiscretization import *
 
 class LSERK134:
     A = np.array([
